@@ -1,11 +1,35 @@
-/** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
+// @ts-check
+
+// This runs in Node.js - Don't use client-side code here (browser APIs, JSX...)
+
+/**
+ * Creating a sidebar enables you to:
+ - create an ordered group of docs
+ - render a sidebar for each doc of that group
+ - provide next/previous navigation
+
+ The sidebars can be generated from the filesystem, or explicitly defined here.
+
+ Create as many sidebars as you want.
+
+ @type {import('@docusaurus/plugin-content-docs').SidebarsConfig}
+ */
 const sidebars = {
-  textbookSidebar: [
+  // By default, Docusaurus generates a sidebar from the docs folder structure
+  tutorialSidebar: [{type: 'autogenerated', dirName: '.'}],
+
+  // But you can create a sidebar manually
+  /*
+  tutorialSidebar: [
+    'intro',
+    'hello',
     {
-      type: 'autogenerated',
-      dirName: '.', // '.' means the sidebar will be generated from the docs folder structure
+      type: 'category',
+      label: 'Tutorial',
+      items: ['tutorial-basics/create-a-document'],
     },
   ],
+   */
 };
 
-module.exports = sidebars;
+export default sidebars;
